Refactor profile edit components to modular sections
Split ProfileEditMode into modular section components for basic info, education, experience, skills, goals, and social links. Added EducationEntryCard and ExperienceEntryCard as reusable form components. Updated types and backend mutation to support optional education field and improved type safety. Improved maintainability and reduced code duplication.
