+++ features/profile/components/edit/index.ts
@@ -1,1 +1,6 @@
 export { BasicInfoSection } from "./BasicInfoSection";
+export { EducationSection } from "./EducationSection";
+export { ExperienceSection } from "./ExperienceSection";
+export { SkillsSection } from "./SkillsSection";
+export { GoalsSection } from "./GoalsSection";
+export { SocialLinksSection } from "./SocialLinksSection";
